[analyzer] Refactoring analyzer config handlers
