+++ analyzer/codechecker_analyzer/analyzers/clangsa/analyzer.py
@@ -122,9 +122,8 @@ class ClangSA(analyzer_base.SourceAnalyzer):
                     analyzer_cmd.extend(cfg)
 
             # Config handler stores which checkers are enabled or disabled.
-            for checker_name, value in config.checks().items():
-                enabled, _ = value
-                if enabled:
+            for checker_name, value in config.checkers.items():
+                if value.enabled:
                     analyzer_cmd.extend(['-Xclang',
                                          '-analyzer-checker=' + checker_name])
                 else:
